AI_UseCase/requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-google-genai
pypdf
faiss-cpu
networkx
sqlalchemy
pandas
python-dotenv
plotly
